+++ kparser2.Analytics.Tests/SettledDivergenceTests.fs
@@ -71,6 +71,7 @@ module private SettledHelpers =
           Value = 1
           Success = "hit"
           CommandNo = commandNo
+          SpellId = None
           MessageId = messageId
           IsProc = false
           ProcValue = 0